Cleanup rule test classes
