+++ pmd-java/src/test/java-resources/net/sourceforge/pmd/lang/java/rule/bestpractices/unittestsshouldusebeforeannotation/BaseTest.java
@@ -1,0 +1,20 @@
+/**
+ * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
+ */
+
+package net.sourceforge.pmd.lang.java.rule.bestpractices.unittestsshouldusebeforeannotation;
+
+import org.junit.After;
+import org.junit.Before;
+
+public class BaseTest {
+    @Before
+    public void setUp() {
+        // A setup code
+    }
+
+    @After
+    public void tearDown() {
+        // A tear down code
+    }
+}
