+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/rule/bestpractices/UnitTestShouldUseBeforeAnnotationTest.java
@@ -4,23 +4,8 @@
 
 package net.sourceforge.pmd.lang.java.rule.bestpractices;
 
-import org.junit.After;
-import org.junit.Before;
-
 import net.sourceforge.pmd.test.PmdRuleTst;
 
 class UnitTestShouldUseBeforeAnnotationTest extends PmdRuleTst {
     // no additional unit tests
-
-    public static class BaseTest {
-        @Before
-        public void setUp() {
-            // A setup code
-        }
-
-        @After
-        public void tearDown() {
-            // A tear down code
-        }
-    }
 }
