+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/rule/codestyle/UnnecessaryImportTest.java
@@ -7,17 +7,5 @@ package net.sourceforge.pmd.lang.java.rule.codestyle;
 import net.sourceforge.pmd.test.PmdRuleTst;
 
 class UnnecessaryImportTest extends PmdRuleTst {
-    // these 2 methods are used for a test case, do not delete
 
-    public static void assertTrue(String message, boolean condition) {
-        if (!condition) {
-            System.out.println(message);
-        }
-    }
-
-    public static void assertSomething(String message, boolean condition) {
-        if (!condition) {
-            System.out.println(message);
-        }
-    }
 }
