+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/rule/multithreading/DontCallThreadRunTest.java
@@ -7,11 +7,4 @@ package net.sourceforge.pmd.lang.java.rule.multithreading;
 import net.sourceforge.pmd.test.PmdRuleTst;
 
 class DontCallThreadRunTest extends PmdRuleTst {
-    // Used by DontCallThreadRun test cases
-    public static class TestThread extends Thread {
-        @Override
-        public void run() {
-            System.out.println("test");
-        }
-    }
 }
