+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/rule/bestpractices/xml/UnitTestShouldUseAfterAnnotation.xml
@@ -76,7 +76,7 @@ public class Foo {
         <description>[java] JUnit4TestShouldUseBeforeAnnotation false positive when overriding setUp #1592</description>
         <expected-problems>0</expected-problems>
         <code><![CDATA[
-import net.sourceforge.pmd.lang.java.rule.bestpractices.UnitTestShouldUseAfterAnnotationTest.BaseTest;
+import net.sourceforge.pmd.lang.java.rule.bestpractices.unittestsshouldusebeforeannotation.BaseTest;
 
 public class AReallyCoolFeatureTest extends BaseTest {
     @Override
