+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/rule/bestpractices/xml/UnitTestShouldUseBeforeAnnotation.xml
@@ -103,7 +103,7 @@ public class Foo {
         <description>[java] JUnit4TestShouldUseBeforeAnnotation false positive when overriding setUp #1592</description>
         <expected-problems>0</expected-problems>
         <code><![CDATA[
-import net.sourceforge.pmd.lang.java.rule.bestpractices.UnitTestShouldUseBeforeAnnotationTest.BaseTest;
+import net.sourceforge.pmd.lang.java.rule.bestpractices.unittestsshouldusebeforeannotation.BaseTest;
 
 public class AReallyCoolFeatureTest extends BaseTest {
     @Override
