+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/rule/codestyle/xml/UnnecessaryImport.xml
@@ -718,7 +718,7 @@ class NPEImport {}
         <expected-problems>0</expected-problems>
         <code><![CDATA[
             import static org.junit.Assert.*;
-            import static net.sourceforge.pmd.lang.java.rule.codestyle.UnnecessaryImportTest.*;
+            import static net.sourceforge.pmd.lang.java.rule.codestyle.unnecessaryimport.AssertContainer.*;
             import static org.junit.Assert.assertTrue;
             // this import is needed for disambiguation - as DuplicateImportsTest
             // defines assertTrue with the same signature, too.
@@ -726,7 +726,7 @@ class NPEImport {}
             public class DuplicateImports {
                 static {
                     assertTrue("", true); // the one from the disambiguation import
-                    assertSomething("", true); // from UnnecessaryImportTest.*
+                    assertSomething("", true); // from AssertContainer.*
                     assertFalse("", true); // from Assert.*
                 }
             }
